Add labels and tags to the ImageVariant model
This sets image labels at various levels throughout the parent Image and
ImageVersion classes.

We intend that these labels will be assembled in the bake plan, and the
image tag prefixes will prefixed there.

Tags are set at the ImageVariant level since tags will be specific to the
OS and target.